[FEAT] #1 - Make recordItem recyclerView
