+++ presentation/src/main/java/com/woowahan/accountbook/ui/itemList/RecordItem.kt
@@ -1,4 +1,94 @@
 package com.woowahan.accountbook.ui.itemList
 
-class RecordItem {
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Divider
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.woowahan.accountbook.ui.theme.*
+
+@Preview
+@Composable
+fun preview() {
+    RecordItem(recordType = "test", paymentType = "test", title = "test", amount = "test")
+}
+
+@Composable
+fun RecordItem(
+    recordType: String,
+    paymentType: String,
+    title: String,
+    amount: String,
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(10.dp, 10.dp, 10.dp, 0.dp)
+    ) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                modifier = Modifier
+                    .width(56.dp)
+                    .clip(CircleShape)
+                    .background(Yellow4)
+                    .padding(10.dp, 5.dp, 10.dp, 5.dp),
+                text = recordType,
+                textAlign = TextAlign.Center,
+                fontSize = 10.sp
+            )
+
+            Spacer(modifier = Modifier.weight(1f))
+            Text(
+                fontSize = 10.sp,
+                color = Purple,
+                text = paymentType
+            )
+        }
+
+        Spacer(modifier = Modifier.height(12.dp))
+
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                fontSize = 14.sp,
+                color = Purple,
+                fontWeight = FontWeight.Bold,
+                text = title
+            )
+            Spacer(modifier = Modifier.weight(1f))
+            Text(
+                fontSize = 14.sp,
+                color = if (amount.startsWith("-")) {
+                    Red
+                } else {
+                    Green6
+                },
+                fontWeight = FontWeight.Bold,
+                text = amount
+            )
+        }
+
+        Spacer(modifier = Modifier.height(10.dp))
+
+        Divider(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(1.dp),
+            color = LightPurple
+        )
+    }
 }
